+++ clikt/src/jvmTest/kotlin/com/github/ajalt/clikt/parsers/ParserTest.kt
@@ -6,11 +6,11 @@ import com.github.ajalt.clikt.parameters.arguments.argument
 import com.github.ajalt.clikt.parameters.arguments.multiple
 import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.testing.TestCommand
-import io.kotest.matchers.shouldBe
 import io.kotest.assertions.throwables.shouldThrow
+import io.kotest.matchers.shouldBe
 import org.junit.Rule
-import kotlin.test.Test
 import org.junit.rules.TemporaryFolder
+import kotlin.test.Test
 
 
 class ParserTest {