+++ samples/helpformat/src/main/kotlin/com/github/ajalt/clikt/samples/helpformat/main.kt
@@ -2,21 +2,21 @@ package com.github.ajalt.clikt.samples.helpformat
 
 import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.core.context
-import com.github.ajalt.clikt.output.HelpFormatter
 import com.github.ajalt.clikt.output.CliktHelpFormatter
+import com.github.ajalt.clikt.output.HelpFormatter
 import com.github.ajalt.clikt.parameters.arguments.argument
 import com.github.ajalt.clikt.parameters.arguments.multiple
 import com.github.ajalt.clikt.parameters.options.flag
 import com.github.ajalt.clikt.parameters.options.option
 
-class ArgparseHelpFormatter: CliktHelpFormatter(
+class ArgparseHelpFormatter : CliktHelpFormatter(
         usageTitle = "usage:",
         optionsTitle = "optional arguments:",
         argumentsTitle = "positional arguments:") {
     override fun formatHelp(prolog: String,
                             epilog: String,
                             parameters: List<HelpFormatter.ParameterHelp>,
-                            programName: String)= buildString {
+                            programName: String) = buildString {
         // argparse prints arguments before options
         addUsage(parameters, programName)
         addProlog(prolog)
@@ -31,6 +31,7 @@ class Echo : CliktCommand(help = "Echo the STRING(s) to standard output") {
     init {
         context { helpFormatter = ArgparseHelpFormatter() }
     }
+
     val suppressNewline by option("-n", help = "do not output the trailing newline").flag()
     val strings by argument(help = "the strings to echo").multiple()